attempt to fix occasional bad value in rx_rate and tx_rate
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@42184 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ reports/hier/usr/lib/python2.7/reports/node/untangle_server.py
@@ -44,7 +44,7 @@ CREATE TABLE reports.server_events (
 CREATE TABLE reports.interface_stat_events (
     time_stamp  TIMESTAMP,
     interface_id INT,
-    rx_rate 	FLOAT,
-    tx_rate 	FLOAT)""")
+    rx_rate 	float8,
+    tx_rate 	float8)""")
 
 reports.engine.register_node(ServerNode())
